+++ tests/Monolog/TraceContextProcessorTest.php
@@ -42,6 +42,7 @@ final class TraceContextProcessorTest extends TestCase
             self::assertArrayHasKey('span_id', $processed->extra);
             self::assertSame($span->getContext()->getTraceId(), $processed->extra['trace_id']);
             self::assertSame($span->getContext()->getSpanId(), $processed->extra['span_id']);
+            self::assertSame(sprintf('%02x', $span->getContext()->getTraceFlags()), $processed->extra['trace_flags']);
         } finally {
             $scope->detach();
             $span->end();